Make build tweaks EPEL compatible
EPEL prefers to be spoon fed some old skool ${SHELL} snippet, rather
than (Lua) macros or scriptlets. It ain't pretty, but it does work.

I also moved the CPU tweak to the relevant place, so it doesn't slow
down the build of the library, which is way less memory hungry.

[skip changelog]